+++ ExpandWorldPrefabs/ObjectsFiltering.cs
@@ -9,58 +9,54 @@ namespace ExpandWorld.Prefab;
 public class ObjectsFiltering
 {
   // Note: Can include the object itself.
-  public static ZDO[] GetNearby(int limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
+  public static ZDO[] GetNearby(int limit, Object[] objects, Vector3 pos, Dictionary<string, string> parameters)
   {
     if (objects.Length == 0) return [];
     var maxRadius = objects.Max(o => o.MaxDistance);
     if (maxRadius > 10000)
     {
       var zdos = ZDOMan.instance.m_objectsByID.Values;
-      return GetObjects(limit, zdos, objects, zdo, parameters);
+      return GetObjects(limit, zdos, objects, pos, parameters);
     }
-    var zdoLists = GetSectorIndices(zdo.m_position, maxRadius);
-    return GetObjects(limit, zdoLists, objects, zdo, parameters);
+    var zdoLists = GetSectorIndices(pos, maxRadius);
+    return GetObjects(limit, zdoLists, objects, pos, parameters);
   }
-  public static ZDO[] GetNearby(int limit, Object objects, ZDO zdo, Dictionary<string, string> parameters)
+  public static ZDO[] GetNearby(int limit, Object objects, Vector3 pos, Dictionary<string, string> parameters)
   {
     if (objects.MaxDistance > 10000)
     {
       var zdos = ZDOMan.instance.m_objectsByID.Values;
-      return GetObjects(limit, zdos, objects, zdo, parameters);
+      return GetObjects(limit, zdos, objects, pos, parameters);
     }
-    var zdoLists = GetSectorIndices(zdo.m_position, objects.MaxDistance);
-    return GetObjects(limit, zdoLists, objects, zdo, parameters);
+    var zdoLists = GetSectorIndices(pos, objects.MaxDistance);
+    return GetObjects(limit, zdoLists, objects, pos, parameters);
   }
-  private static ZDO[] GetObjects(int limit, List<List<ZDO>> zdoLists, Object objects, ZDO zdo, Dictionary<string, string> parameters)
+  private static ZDO[] GetObjects(int limit, List<List<ZDO>> zdoLists, Object objects, Vector3 pos, Dictionary<string, string> parameters)
   {
-    var pos = zdo.m_position;
     var zm = ZDOMan.instance;
     var query = zdoLists.SelectMany(z => z).Where(z => objects.IsValid(z, pos, parameters));
     if (limit > 0)
       query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
     return query.ToArray();
   }
-  private static ZDO[] GetObjects(int limit, Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object objects, ZDO zdo, Dictionary<string, string> parameters)
+  private static ZDO[] GetObjects(int limit, Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object objects, Vector3 pos, Dictionary<string, string> parameters)
   {
-    var pos = zdo.m_position;
     var zm = ZDOMan.instance;
     var query = zdos.Where(z => objects.IsValid(z, pos, parameters));
     if (limit > 0)
       query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
     return query.ToArray();
   }
-  private static ZDO[] GetObjects(int limit, List<List<ZDO>> zdoLists, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
+  private static ZDO[] GetObjects(int limit, List<List<ZDO>> zdoLists, Object[] objects, Vector3 pos, Dictionary<string, string> parameters)
   {
-    var pos = zdo.m_position;
     var zm = ZDOMan.instance;
     var query = zdoLists.SelectMany(z => z).Where(z => objects.Any(o => o.IsValid(z, pos, parameters)));
     if (limit > 0)
       query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
     return query.ToArray();
   }
-  private static ZDO[] GetObjects(int limit, Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
+  private static ZDO[] GetObjects(int limit, Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object[] objects, Vector3 pos, Dictionary<string, string> parameters)
   {
-    var pos = zdo.m_position;
     var zm = ZDOMan.instance;
     var query = zdos.Where(z => objects.Any(o => o.IsValid(z, pos, parameters)));
     if (limit > 0)