+++ ExpandWorldPrefabs/service/CommandManager.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+using Data;
 using UnityEngine;
 
 namespace Service;
@@ -62,7 +63,7 @@ public class CommandManager
       }
     }
   }
-  public static void Run(IEnumerable<string> commands, PlayerInfo[] players)
+  public static void Run(IEnumerable<string> commands, List<PlayerInfo> players)
   {
     foreach (var peer in players)
       Run(commands, peer);
@@ -87,58 +88,9 @@ public class CommandManager
       // Single negative number would get handled as expression.
       var sub = expression.Substring(1);
       if (!sub.Contains('*') && !sub.Contains('/') && !sub.Contains('+') && !sub.Contains('-')) continue;
-      var value = Evaluate(expression);
+      var value = Calculator.EvaluateFloat(expression) ?? 0f;
       cmd = cmd.Replace(expression, value.ToString("0.#####", NumberFormatInfo.InvariantInfo));
     }
     return cmd;
-  }
-  private static float Evaluate(string expression)
-  {
-    var mult = expression.Split('*');
-    if (mult.Length > 1)
-    {
-      var sum = 1f;
-      foreach (var m in mult) sum *= Evaluate(m);
-      return sum;
-    }
-    var div = expression.Split('/');
-    if (div.Length > 1)
-    {
-      var sum = Evaluate(div[0]);
-      for (var i = 1; i < div.Length; ++i) sum /= Evaluate(div[i]);
-      return sum;
-    }
-    var plus = expression.Split('+');
-    if (plus.Length > 1)
-    {
-      var sum = 0f;
-      foreach (var p in plus) sum += Evaluate(p);
-      return sum;
-    }
-    var minus = expression.Split('-');
-    // Negative numbers get split as well, so check for actual parts.
-    if (minus.Where(s => s != "").Count() > 1)
-    {
-      float? sum = null;
-      for (var i = 0; i < minus.Length; ++i)
-      {
-        if (minus[i] == "" && i + 1 < minus.Length)
-        {
-          minus[i + 1] = "-" + minus[i + 1];
-          continue;
-        }
-        if (sum == null) sum = Evaluate(minus[i]);
-        else sum -= Evaluate(minus[i]);
-      }
-      return sum ?? 0;
-    }
-    try
-    {
-      return float.Parse(expression.Trim(), NumberFormatInfo.InvariantInfo);
-    }
-    catch
-    {
-      throw new Exception($"Failed to parse expression: {expression}");
-    }
   }
 }